requirements.txt
pyyaml
pyroute2
protobuf ~= 3.20.1
grpcio
grpcio-tools